desktop: Various cleanups
