+++ src/desktop/mod.rs
@@ -1,7 +1,9 @@
 // TODO: Remove - but for now, this makes sure these files are not completely highlighted with warnings
 #![allow(missing_docs, clippy::all)]
+mod output;
 mod popup;
 mod space;
+pub mod utils;
 mod window;
 
 pub use self::popup::*;
